Move IdMap and tests to new file
